+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Development
 
 - Fix error when a registered non-admin user is looking at `/organization` and would see a technical group in the list of orgs (which they are not authorized to see).
+- Fix error where password reset links were rejected.
 
 ## [0.2.4](https://github.com/berlinonline/ckanext-berlinauth/releases/tag/0.2.4)